+++ cypress/e2e/ask_custom_element/spec.cy.ts
@@ -7,9 +7,10 @@ describe('Ask Custom Element', () => {
 
   it('should send element props to the backend', () => {
     cy.get('.step').should('have.length', 1);
+    cy.get('#ticket-submit').should('be.disabled');
     cy.get('#summary').type('Bug fix');
     cy.get('#description').type('Detailed description');
-    cy.get('#ticket-submit').click();
+    cy.get('#ticket-submit').should('not.be.disabled').click();
     cy.get('.step').should('have.length', 3);
     cy.get('.step').eq(2).should('contain', 'Bug fix');
   });